backend/requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
psycopg2-binary>=2.9,<3.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0

